+++ rosomaxa/src/algorithms/gsom/contraction.rs
@@ -27,14 +27,14 @@ where
     // find nodes which should be removed
     let removed = network
         .get_nodes()
-        .map(|node| node.read().unwrap().coordinate)
+        .map(|node| node.coordinate)
         .filter(|coord| coord.0 % x_decim == 0 || coord.1 % y_decim == 0)
         .collect::<Vec<_>>();
 
     // remove nodes with given coordinates, but keep track of their data
     let data = removed.iter().fold(Vec::new(), |mut data, coordinate| {
-        let node = network.find(coordinate).unwrap();
-        data.extend(node.write().unwrap().storage.drain(0..));
+        let node = network.get_mut(coordinate).unwrap();
+        data.extend(node.storage.drain(0..));
         network.remove(coordinate);
 
         data
@@ -43,10 +43,10 @@ where
     // detect what was deleted and shift coordinates of all affected nodes to retain connectivity
     // shift not only to the right/top, but also to the left/bottom to keep center around (0, 0)
     network.remap(&|Coordinate(x, y), node| {
+        let mut node = node;
         let x = x + get_offset(x, (x_min, x_max), x_decim);
         let y = y + get_offset(y, (y_min, y_max), y_decim);
-
-        node.write().unwrap().coordinate = Coordinate(x, y);
+        node.coordinate = Coordinate(x, y);
 
         node
     });